chore: interval jitter calculation now uses native u128

+++ src/dht.rs
@@ -154,7 +154,9 @@ impl DhtActor {
             self.reset().await?;
 
             let jitter = if self.config.max_retry_jitter() > Duration::ZERO {
-                Duration::from_nanos(rand::random::<u64>() % self.config.max_retry_jitter().as_nanos() as u64)
+                Duration::from_nanos(
+                    (rand::random::<u128>() % self.config.max_retry_jitter().as_nanos()) as u64,
+                )
             } else {
                 Duration::ZERO
             };
